optimize summary logic

+++ lib/analyzer/ershou_analyzer.py
@@ -15,9 +15,9 @@ class AreaErShou(object):
         "总价平均(万)," + \
         "面积平均(平米)," + \
         "建成年份平均\n"
-    def __init__(self, date, distrcit, area, quantity, unit_price_avrg, total_price_avrg, mianji_avrg, by_avrg):
+    def __init__(self, date, district, area, quantity, unit_price_avrg, total_price_avrg, mianji_avrg, by_avrg):
         self.date = date
-        self.distrcit = distrcit
+        self.district = district
         self.area = area
         self.quantity = quantity
         self.up_avrg = unit_price_avrg # unit_price_avrg
@@ -27,7 +27,7 @@ class AreaErShou(object):
 
     def text(self):
         return self.date + "," + \
-            self.distrcit + "," + \
+            self.district + "," + \
             self.area + "," + \
             str(self.quantity) + "," + \
             str(self.up_avrg) + "," + \
@@ -39,13 +39,13 @@ class AreaErShou(object):
 class ErShouAnalyzer(object):
     def __init__(self):
         self.city_pinyin_name = None
-        self.today_path = None
+        self.summary_path = None
         self.district_areas_house_info_dict = dict()
         self.summary_list = list()
 
-    def set_base_info(self, today_path, city_pinyin_name):
+    def set_base_info(self, city_pinyin_name, summary_path):
         self.city_pinyin_name = city_pinyin_name
-        self.today_path = today_path
+        self.summary_path = summary_path
 
     def add_area_houses_info_to_dict(self,
                                      district_pinyin_name,
@@ -57,10 +57,13 @@ class ErShouAnalyzer(object):
         area_houses_dict = self.district_areas_house_info_dict[district_pinyin_name]
         area_houses_dict[area_pinyin_name] = ershou_houses_list
 
-        print("add {0} {1} houses to {2}".format(
-            area_pinyin_cn_name_dict[area_pinyin_name],
-            len(ershou_houses_list),
-            district_pinyin_cn_name_dict[district_pinyin_name]))
+        try:
+            print("add {0} {1} houses to {2}".format(
+                area_pinyin_cn_name_dict[area_pinyin_name],
+                len(ershou_houses_list),
+                district_pinyin_cn_name_dict[district_pinyin_name]))
+        except Exception as e:
+            print("error: ", e)
 
     def average_calculator(self, ershous_list):
         unit_price = 0
@@ -133,33 +136,42 @@ class ErShouAnalyzer(object):
         mj_avrg_sum = 0
         by_avrg_sum = 0
 
-        csv_file = self.today_path + "/{0}_summary.csv".format(self.city_pinyin_name)
-        print(csv_file)
+        for summary_item in self.summary_list:
+            summary_item_sum += 1
+            quantity_sum += summary_item.quantity
+            up_avrg_sum += summary_item.up_avrg
+            tp_avrg_sum += summary_item.tp_avrg
+            mj_avrg_sum += summary_item.mianji_avrg
+            by_avrg_sum += summary_item.by_avrg
 
-        with open(csv_file, "w") as f:
-            f.write(AreaErShou.summary_title) # write first row(title) of csv file
-            for summary_item in self.summary_list:
+            area_csv_file = self.summary_path + "/{0}_{1}_summary.csv".format(summary_item.district, summary_item.area)
+            print(area_csv_file)
+
+            if not os.path.isfile(area_csv_file):
+                with open(area_csv_file, "w") as f:
+                    f.write(AreaErShou.summary_title) # write first row(title) of csv file
+
+            with open(area_csv_file, "a") as f:
                 f.write(summary_item.text())
 
-                summary_item_sum += 1
-                quantity_sum += summary_item.quantity
-                up_avrg_sum += summary_item.up_avrg
-                tp_avrg_sum += summary_item.tp_avrg
-                mj_avrg_sum += summary_item.mianji_avrg
-                by_avrg_sum += summary_item.by_avrg
+        csv_file = self.summary_path + "/{0}_summary.csv".format(self.city_pinyin_name)
+        print(csv_file)
+        if not os.path.isfile(csv_file):
+            with open(csv_file, "w") as f:
+                f.write(AreaErShou.summary_title) # write first row(title) of csv file
 
+        with open(csv_file, "a") as f:
             up_avrg_sum_avrg = up_avrg_sum/summary_item_sum
             tp_avrg_sum_avrg = tp_avrg_sum/summary_item_sum
             mj_avrg_sum_avrg = mj_avrg_sum/summary_item_sum
             by_avrg_sum_avrg = by_avrg_sum/summary_item_sum
 
-            f.write("N/A," + self.city_pinyin_name + ",N/A," + \
+            f.write(self.summary_list[0].date + "," + self.city_pinyin_name + ",N/A," + \
                     str(quantity_sum) + "," + \
                     str(up_avrg_sum_avrg) + "," + \
                     str(tp_avrg_sum_avrg) + "," + \
                     str(mj_avrg_sum_avrg) + "," + \
                     str(by_avrg_sum_avrg) + "\n")
-
 
 
 if __name__ == "__main__":
@@ -177,33 +189,37 @@ if __name__ == "__main__":
         ershou_list = list()
 
         for i in range(1, len(csv_item_list)):
-            ershou_item = csv_item_list[i]
+            try:
+                ershou_item = csv_item_list[i]
                 # 作为对象保存
-            ershou = ErShou(ershou_item[0], ershou_item[1], ershou_item[2], ershou_item[3],
+                ershou = ErShou(ershou_item[0], ershou_item[1], ershou_item[2], ershou_item[3],
                             ershou_item[4], ershou_item[5], ershou_item[6], ershou_item[7],
                             ershou_item[8])
 
-            ershou.set_value(ershou_item[0], ershou_item[1], ershou_item[2], ershou_item[3],
+                ershou.set_value(ershou_item[0], ershou_item[1], ershou_item[2], ershou_item[3],
                             ershou_item[4], ershou_item[5], ershou_item[6], ershou_item[7],
                             ershou_item[8], ershou_item[9], ershou_item[10], ershou_item[11],
                             ershou_item[12], ershou_item[13], ershou_item[14])
 
-            ershou_list.append(ershou)
+                ershou_list.append(ershou)
+            except Exception as e:
+                print("current item: ", ershou_item)
 
         return ershou_list
 
-    city_pinyin_name = 'sh'
+    city_pinyin_name = 'su'
     districts_pinyin_names = get_districts(city_pinyin_name)
     for district_pinyin_name in districts_pinyin_names:
         area_pinyin_names_of_district = get_areas(city_pinyin_name, district_pinyin_name)
 
-    today_path = path.DATA_PATH + "/" + SPIDER_NAME + "/ershou/" + city_pinyin_name + "/20230924"
-    print("today csv data path: ", today_path)
+    today_path = path.DATA_PATH + "/" + SPIDER_NAME + "/ershou/" + city_pinyin_name + "/20231015"
+    summary_path = path.create_summary_path("{0}/ershou".format(SPIDER_NAME), city_pinyin_name)
+    print("today_path: ", today_path)
 
     csv_lists = os.listdir(today_path)
 
     ershou_analyzer = ErShouAnalyzer()
-    ershou_analyzer.set_base_info(today_path, city_pinyin_name)
+    ershou_analyzer.set_base_info(city_pinyin_name, summary_path)
     for csv_file in csv_lists:
         if csv_file == city_pinyin_name + "_summary.csv":
             continue
